pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devtoys-tg-bot"
version = "0.1.0"
description = "Self-hostable Telegram bot inspired by DevToys"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DevToys Telegram Bot Team", email = "ops@example.com"}
]
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: AsyncIO",
    "Topic :: Communications :: Chat"
]
dependencies = [
    "aiogram>=3.4,<4.0",
    "structlog>=24.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "orjson>=3.9",
    "redis>=5.0",
    "python-ulid>=2.2",
    "pillow>=10.0",
    "qrcode[pil]>=7.4",
    "python-dateutil>=2.8",
    "pendulum>=3.0",
    "regex>=2023.12",
    "httpx>=0.25",
    "aiofiles>=23.0",
    "uvloop>=0.19; platform_system != \"Windows\""
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.11",
    "ruff>=0.3",
    "black>=24.0",
    "types-PyYAML",
    "types-python-dateutil"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "UP", "B", "N", "ASYNC"]
ignore = ["B008"]
src = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
source = ["src"]
branch = true
